+++ src/components/PerfumeLab/NoteSelectionContent.tsx
