+++ src/components/PerfumeLab/NoteSelectionSection.tsx
