+++ src/components/PerfumeLab/PerfumeNoteInfo.tsx
