+++ src/components/PerfumeLab/Result/ContentBox.tsx
