+++ src/components/PerfumeLab/Result/PerfumeCard.tsx
