+++ src/components/PerfumeLab/Result/PerfumeResultDisplay.tsx
