+++ src/components/PerfumeLab/Result/ResultContent.tsx
