+++ src/components/PerfumeLab/Slider.tsx
